+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next version
 
+- Add `SmartKandelSeeder` contract and deployment on Blast
+
 # 2.2.3
 
 - fix: `KandelSeeder` address on Polygon Mumbai reverted (mistakenly changed to a Blast address in v2.2.1)